lessons1-Task4 changing the type of variable

+++ src/main/kotlin/Lessons1-Task4.kt
@@ -1,8 +1,8 @@
-fun main(){
+fun main() {
 
-    val lenght: Long = 40868600000
+    val lenght: Long = 40868600000L
     val age: Byte = 27
-    val partOfDay: Double = 0.075
+    val partOfDay: Float = 0.075F
     val second: Short = 6480
     val partOfYear: String = "2.0547945205479453E-4"
     val apogee: Int = 327000
